+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2023-03-30
+
+### Added
+
+- Registration will be every `n` minutes to be sure that module is registered
+- Registration logic
+
+### Changed
+
+- Add submodule to get user by id request 
+
 ## [1.0.4] - 2023-03-27
 
 ### Added